+++ docs/conf.py
@@ -1,0 +1,103 @@
+# Configuration file for the Sphinx documentation builder.
+#
+# This file only contains a selection of the most common options. For a full
+# list see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# -- Path setup --------------------------------------------------------------
+
+# If extensions (or modules to document with autodoc) are in another directory,
+# add these directories to sys.path here. If the directory is relative to the
+# documentation root, use os.path.abspath to make it absolute, like shown here.
+#
+# import os
+# import sys
+#
+# sys.path.insert(0, os.path.abspath("../."))
+
+import sphinx_rtd_theme
+
+import pillow_heif
+
+# -- Project information -----------------------------------------------------
+
+project = "pillow-heif"
+copyright = "2021-2022, Alexander Piskun and Contributors"
+author = "Alexander Piskun and Contributors"
+
+# The short X.Y version.
+version = pillow_heif.__version__
+# The full version, including alpha/beta/rc tags
+release = pillow_heif.__version__
+
+# -- General configuration ---------------------------------------------------
+
+# If your documentation needs a minimal Sphinx version, state it here.
+#
+needs_sphinx = "4.4"
+
+# Add any Sphinx extension module names here, as strings. They can be
+# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
+# ones.
+extensions = [
+    "sphinx_issues",
+    "sphinx.ext.autodoc",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.viewcode",
+]
+
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3", None),
+    "pillow": ("https://pillow.readthedocs.io/en/stable", None),
+}
+
+# The suffix(es) of source filenames.
+# You can specify multiple suffix as a list of string:
+# source_suffix = ['.rst', '.md']
+source_suffix = ".rst"
+
+# List of patterns, relative to source directory, that match files and
+# directories to ignore when looking for source files.
+# This pattern also affects html_static_path and html_extra_path.
+exclude_patterns = ["_build"]
+
+# The name of the Pygments (syntax highlighting) style to use.
+pygments_style = "sphinx"
+
+# If true, `todos` produce output, else they produce nothing.
+todo_include_todos = False
+
+# If true, Sphinx will warn about all references where the target cannot be found.
+# Default is False. You can activate this mode temporarily using the -n command-line
+# switch.
+nitpicky = True
+
+# -- Options for HTML output -------------------------------------------------
+
+# The theme to use for HTML and HTML Help pages.  See the documentation for
+# a list of builtin themes.
+#
+html_theme = "sphinx_rtd_theme"
+html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+
+# Add any paths that contain custom static files (such as style sheets) here,
+# relative to this directory. They are copied after the builtin static files,
+# so a file named "default.css" will overwrite the builtin "default.css".
+html_static_path = ["resources"]
+
+
+def setup(app):
+    app.add_js_file("js/script.js")
+    app.add_css_file("css/styles.css")
+    app.add_css_file("css/dark.css")
+    app.add_css_file("css/light.css")
+
+
+# GitHub repo for sphinx-issues
+issues_github_path = "bigcat88/pillow_heif"
+
+# 'short' – Suppress the leading module names of the typehints (ex. io.StringIO -> StringIO)
+autodoc_typehints_format = "short"
+
+# Do not sort members by alphabet.
+autodoc_member_order = "bysource"