+++ pillow_heif/__init__.py
@@ -17,7 +17,6 @@ from .constants import heif_filetype_no  # DEPRECATED
 from .constants import heif_filetype_yes_supported  # DEPRECATED
 from .constants import heif_filetype_yes_unsupported  # DEPRECATED
 from .constants import (
-    HeifBrand,
     HeifChannel,
     HeifChroma,
     HeifColorProfileType,