+++ pillow_heif/misc.py
@@ -43,12 +43,13 @@ def reset_orientation(info: dict) -> Union[int, None]:
 
 def get_file_mimetype(fp) -> str:
     """
-    Wrapper around `libheif.get_file_mimetype`.
+    Wrapper around `libheif.get_file_mimetype`
 
     :param fp: A filename (string), pathlib.Path object, file object or bytes.
        The file object must implement ``file.read``, ``file.seek`` and ``file.tell`` methods,
        and be opened in binary mode.
-    :returns: string with `image/*`. If the format could not be detected, an empty string is returned.
+    :returns: "image/heic", "image/heif", "image/heic-sequence",
+        "image/heif-sequence", "image/avif" or "image/avif-sequence"
     """
     __data = _get_bytes(fp, 50)
     return ffi.string(lib.heif_get_file_mime_type(__data, len(__data))).decode()